Refactor calendar module into `eventScheduler` and `calendarView`:
- Moved scheduling logic to `eventScheduler.py` under a new `Scheduler` class.
- Extracted event viewing functionality into `calendarView.py`, utilizing the `CalendarView` class.
- Updated `tokenize.py` to reference `eventScheduler` and `calendarView`.
- Removed redundant function `convertListToText` from `dbUtils.py`.
